# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.3.0
  description: SOC OS - SaaS Security Platform API (V30/V50)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List security events
      parameters:
        - name: action
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: tactic
          in: query
          schema:
            type: string
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of security events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SecurityEvent"
    post:
      operationId: processEvent
      tags: [events]
      summary: Process a security event through the full SOC pipeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "200":
          description: Pipeline result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineResult"
...
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get a single security event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Security event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityEvent"
        "404":
          description: Not found
  /events/{id}/status:
    patch:
      operationId: updateEventStatus
      tags: [events]
      summary: Update alert lifecycle status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventStatusBody"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityEvent"
        "404":
          description: Not found
...
components:
  schemas:
...
    SecurityEvent:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: ["integer", "null"]
        event:
          type: string
        score:
          type: integer
        action:
          type: string
          enum: [ALLOW, WARN, ISOLATE, PATCHED]
        status:
          type: string
          enum: [NEW, ACKNOWLEDGED, INVESTIGATING, QUARANTINED, RESOLVED]
        tactic:
          type: ["string", "null"]
        technique:
          type: ["string", "null"]
        techniqueId:
          type: ["string", "null"]
        velocityFlag:
          type: boolean
        correlationId:
          type: ["integer", "null"]
        nodeId:
          type: string
        hash:
          type: string
        prevHash:
          type: string
        timestamp:
          type: string
          format: date-time
        cpuUsage:
          type: ["integer", "null"]
        memUsage:
          type: ["integer", "null"]
      required:
        - id
        - event
        - score
        - action
        - status
        - velocityFlag
        - nodeId
        - hash
        - prevHash
        - timestamp
    EventInput:
      type: object
      properties:
        event:
          type: string
          minLength: 1
        tenantId:
          type: ["integer", "null"]
        cpuUsage:
          type: ["integer", "null"]
        memUsage:
          type: ["integer", "null"]
        managedResourceKey:
          type: string
          pattern: "^[a-zA-Z0-9][a-zA-Z0-9._:-]{0,127}$"
        observedManagedState:
          type: object
          additionalProperties: true
      required: [event]
    PipelineResult:
      type: object
      properties:
        event:
          $ref: "#/components/schemas/SecurityEvent"
        rulesMatched:
          type: integer
        scoreBoost:
          type: integer
        correlation:
          nullable: true
          oneOf:
            - $ref: "#/components/schemas/Correlation"
            - type: "null"
        pipelineStages:
          type: array
          items:
            type: string
        selfHealing:
          nullable: true
          oneOf:
            - $ref: "#/components/schemas/SelfHealingAction"
            - type: "null"
      required: [event, rulesMatched, scoreBoost, pipelineStages]
...
    SelfHealingAction:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: ["integer", "null"]
        eventId:
          type: ["integer", "null"]
        restorePointId:
          type: ["integer", "null"]
        resourceKey:
          type: string
        location:
          type: string
        mode:
          type: string
          enum: [PREVIEW, APPLY, AUTO]
        status:
          type: string
          enum: [READY, RESTORED, NO_RESTORE_POINT, REJECTED, FAILED]
        previousStateHash:
          type: ["string", "null"]
        restoredStateHash:
          type: ["string", "null"]
        message:
          type: string
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
      required: [id, resourceKey, location, mode, status, message, createdAt]
...
    UpdateEventStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [ACKNOWLEDGED, INVESTIGATING, RESOLVED]
      required: [status]
...
    Correlation:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: ["integer", "null"]
        threatType:
          type: string
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        confidence:
          type: number
        summary:
          type: string
        eventIds:
          type: string
        resolvedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, threatType, severity, confidence, summary, eventIds, createdAt]
...